Update Rust deps (#2332)
* Temporarily disable hakari

* Upgrade compatible deps except Chrono

* Update semver-incompatible crates

* Re-enable hakari

* Update licenses & cargo-deny

* Fix new clippy lints

* Update to latest Rust

+++ rslib/src/decks/schema11.rs
@@ -371,7 +371,7 @@ impl From<Deck> for DeckSchema11 {
                 } else {
                     Some(filt.delays.clone())
                 },
-                preview_delay: filt.preview_delay as u32,
+                preview_delay: filt.preview_delay,
                 common: deck.into(),
             }),
         }
